working solution 1
