+++ src/test/java/com/adobe/campaign/tests/integro/phased/events/data/simple1/PhasedChild2.java
@@ -9,7 +9,7 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.adobe.campaign.tests.integro.phased.events.data;
+package com.adobe.campaign.tests.integro.phased.events.data.simple1;
 
 import com.adobe.campaign.tests.integro.phased.PhasedDataProvider;
 import com.adobe.campaign.tests.integro.phased.PhasedTest;
@@ -20,15 +20,21 @@ import org.testng.annotations.Test;
 //@Test(groups = "aaa", dataProvider = "MUTATIONAL", dataProviderClass = PhasedDataProvider.class)
 @PhasedTest
 @Test(groups = "aaa")
-public class PhasedChild1 extends PhasedParent {
+public class PhasedChild2 extends PhasedParent {
 
     public void step1(String phaseGroup) {
-        System.out.println("Executing step1 with"+phaseGroup);
+        System.out.println("PG2 : Executing step1 with"+phaseGroup);
         Assert.assertEquals(1,1);
     }
 
     public void step2(String phaseGroup) {
-        System.out.println("Executing step2 with"+phaseGroup);
+        System.out.println("PG2 : Executing step2 with"+phaseGroup);
+
+        Assert.assertEquals(1,1);
+    }
+
+    public void step3(String phaseGroup) {
+        System.out.println("PG2 : Executing step3 with"+phaseGroup);
 
         Assert.assertEquals(1,1);
     }
